Move pdgw code into r2harvest.cpp, use R_LOG_INFO and r2ghidra.write.* config
