+++ src/Makefile
@@ -13,6 +13,7 @@ R2GHIDRA_SRCS+=CodeXMLParse.cpp
 R2GHIDRA_SRCS+=ArchMap.cpp
 R2GHIDRA_SRCS+=R2PrintC.cpp
 R2GHIDRA_SRCS+=RCoreMutex.cpp
+R2GHIDRA_SRCS+=r2harvest.cpp
 
 R2GHIDRA_SRCS+=SleighAnalValue.cpp
 R2GHIDRA_SRCS+=SleighAsm.cpp
